Update process docs to reflect recent changes to release process
Ask discussed the new plan is to have the version stored on
`main` reflect the next upcoming, but unreleased, version of
emscripten.   This matches, for example, llvm where the git
users see version N+.

The hope is that this means that we can tag that exact commit
that goes into an emsdk releases rather than the next commit,
and it should avoid tree closures.